Finishes valid moves and threatened squares, fixes many bugs, updates tests, organizes code
My plan has changed a bit. To determine whether the king is in check, I had the idea for a list of "threatened" squares. The implementation works great so far.

The bugfixes mostly related to the coordinate system and logical errors.

I rewrote all the getValidMovesAndThreats methods to fix them and make them neater.

The next step is finishing leavesKingInCheck, then allowing moves to actually be made.

Estimated time spent: 12 hours

Everything is going smoothly!